health damage and particle effects
ive added health, a system that does damage to the car when it bumps into stuff and multiplies it by the cars speed and the speed of the object it hits, and some temporary particle effects for half health and three quarters of health. tweaked the enemy ai and added the subaru model.